Used template method pattern
