+++ src/main/java/edu/kis/powp/jobs2d/command/transformations/ITransformation.java
@@ -1,0 +1,9 @@
+package edu.kis.powp.jobs2d.command.transformations;
+
+import edu.kis.powp.jobs2d.command.CommandCoordinatesVisitor;
+import edu.kis.powp.jobs2d.command.ICompoundCommand;
+
+public interface ITransformation {
+
+	public ICompoundCommand transform(CommandCoordinatesVisitor visitor);
+}
