+++ src/main/java/edu/kis/powp/jobs2d/command/transformations/Move.java
@@ -2,31 +2,22 @@ package edu.kis.powp.jobs2d.command.transformations;
 
 import java.util.List;
 
-import edu.kis.powp.jobs2d.command.CommandCoordinatesVisitor;
-import edu.kis.powp.jobs2d.command.DriverCommand;
-import edu.kis.powp.jobs2d.command.ICompoundCommand;
-import edu.kis.powp.jobs2d.command.ImmutableComplexCommand;
 import edu.kis.powp.jobs2d.command.line.Line2d;
 
-public class Move implements Transformation {
+public class Move extends Transformation {
 	
 	private int x;
 	
 	private int y;
 	
 	public Move(int x, int y) {
+		super();
 		this.x = x;
 		this.y = y;
 	}
 
 	@Override
-	public ICompoundCommand transform(CommandCoordinatesVisitor visitor) {
-		List<Line2d> lines = visitor.getAllCommandsCoordinates();
-
-		TransformationManager.moveToPosition(lines, x, y);
-
-		List<DriverCommand> commandList = TransformationManager.buildCommandList(lines);
-
-		return new ImmutableComplexCommand(commandList);
+	protected void properTransformation(List<Line2d> lines) {
+		moveToPosition(lines, x, y);
 	}
 }
