+++ src/main/java/edu/kis/powp/jobs2d/command/transformations/Rotation.java
@@ -2,30 +2,24 @@ package edu.kis.powp.jobs2d.command.transformations;
 
 import java.util.List;
 
-import edu.kis.powp.jobs2d.command.CommandCoordinatesVisitor;
-import edu.kis.powp.jobs2d.command.DriverCommand;
-import edu.kis.powp.jobs2d.command.ICompoundCommand;
-import edu.kis.powp.jobs2d.command.ImmutableComplexCommand;
 import edu.kis.powp.jobs2d.command.line.Line2d;
 
-public class Rotation implements Transformation {
+public class Rotation extends Transformation {
 
 	private int rotateAngle;
 	
 	public Rotation(int rotateAngle) {
+		super();
 		this.rotateAngle = rotateAngle;
 	}
 	
 	@Override
-	public ICompoundCommand transform(CommandCoordinatesVisitor visitor) {
-		
-		List<Line2d> lines = visitor.getAllCommandsCoordinates();
-
+	protected void properTransformation(List<Line2d> lines) {
 		int x = lines.get(0).getStartPosX();
 		int y = lines.get(0).getStartPosY();
 
 		// Move drawing to position 0,0
-		TransformationManager.moveToPosition(lines, -x, -y);
+		moveToPosition(lines, -x, -y);
 
 
 		int startX, startY, endX, endY;
@@ -45,10 +39,6 @@ public class Rotation implements Transformation {
 		}
 
 		// move back to original position
-		TransformationManager.moveToPosition(lines, x, y);
-
-		List<DriverCommand> commandList = TransformationManager.buildCommandList(lines);
-
-		return new ImmutableComplexCommand(commandList);
+		moveToPosition(lines, x, y);
 	}
 }
